Improves validation and secures user operations
Enforces data ownership by utilizing authenticated user IDs for profile updates and deletions instead of relying on request parameters.

Introduces explicit validation checks and standardized error responses across controllers to ensure data integrity.

Refactors the user update service to use database transactions for more robust state management.

Updates API documentation to reflect authentication requirements and changes in request body structures.

Secure user operations and improve data validation

Enforces data ownership by utilizing authenticated user IDs for profile updates and deletions instead of relying on request parameters.

Introduces explicit validation checks across controllers to ensure data integrity and provide standardized error responses.

Refactors the database schema and service layer to utilize transactions and a new answer grouping structure for more robust state management.

Optimizes AI prompts for stricter JSON output and adds an analysis service for processing career interest results.

Updates API documentation to reflect authentication requirements and changes in request structures.

+++ backend/src/question/dto/question_create.dto.js
@@ -1,5 +1,5 @@
 import { z } from "zod";
 
 export const QuestionCreateDto = z.object({
-  user_id: z.number().int(),
+  user_id: z.number(),
 });
